Refactor getFirstQueryClientIdFromBlocks to utils.js
