+++ packages/block-library/src/comments-query/edit/query-setup.js
@@ -9,6 +9,11 @@ import {
 	__experimentalBlockPatternSetup as BlockPatternSetup,
 } from '@wordpress/block-editor';
 
+/**
+ * Internal dependencies
+ */
+import { getFirstQueryClientIdFromBlocks } from '../utils';
+
 const QueryPlaceholder = () => {
 	const blockProps = useBlockProps();
 
@@ -16,26 +21,6 @@ const QueryPlaceholder = () => {
 	return (
 		<div { ...blockProps }> QueryPlacholder: Actually implement this </div>
 	);
-};
-
-/**
- * Recurses over a list of blocks and returns the clientId
- * of the first found Comments Query Loop block.
- *
- * @param {WPBlock[]} blocks The list of blocks to look through.
- * @return {string=} The clientId.
- */
-export const getFirstQueryClientIdFromBlocks = ( blocks ) => {
-	const blocksQueue = [ ...blocks ];
-	while ( blocksQueue.length > 0 ) {
-		const block = blocksQueue.shift();
-		if ( block.name === 'core/comments-query' ) {
-			return block.clientId;
-		}
-		block.innerBlocks?.forEach( ( innerBlock ) => {
-			blocksQueue.push( innerBlock );
-		} );
-	}
 };
 
 export const QueryPatternSetup = ( props ) => {
